+++ src/pages/AddDataProduct/AddDataProduct.tsx
@@ -5,6 +5,7 @@ import { Grid, Paper } from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
 import { storageObject } from '@ska-telescope/ska-gui-local-storage';
 import {
+  AlertColorTypes,
   Button,
   ButtonColorTypes,
   ButtonVariantTypes,
@@ -15,6 +16,7 @@ import {
 } from '@ska-telescope/ska-gui-components';
 import PageBanner from '../../components/layout/pageBanner/PageBanner';
 import { IMAGE_SIZE_UNITS, NAV, STATUS_OK } from '../../utils/constants';
+import Alert from '../../components/alerts/standardAlert/StandardAlert';
 import HelpPanel from '../../components/info/helpPanel/helpPanel';
 import Proposal from '../../utils/types/proposal';
 import FieldWrapper from '../../components/wrappers/fieldWrapper/FieldWrapper';
@@ -22,9 +24,6 @@ import ImageWeightingField from '../../components/fields/imageWeighting/imageWei
 import { SensCalcResults } from '../../utils/types/sensCalcResults';
 import DataProductSDP from '../../utils/types/dataProduct';
 import Observation from '../../utils/types/observation';
-
-// TODO : Add documentation page specifically for Adding a Data product
-// TODO : Replace individual tick-boxes with a mapping
 
 const BACK_PAGE = 7;
 const PAGE = 13;
@@ -42,7 +41,6 @@ export default function AddDataProduct() {
   const [dp2, setDP2] = React.useState(false);
   const [dp3, setDP3] = React.useState(false);
   const [dp4, setDP4] = React.useState(false);
-  const [dp5, setDP5] = React.useState(false);
   const [imageSizeValue, setImageSizeValue] = React.useState(0);
   const [imageSizeUnits, setImageSizeUnits] = React.useState(IMAGE_SIZE_UNITS.DEGREES);
   const [pixelSizeValue, setPixelSizeValue] = React.useState(0);
@@ -155,7 +153,6 @@ export default function AddDataProduct() {
         {tickElement(2, dp2, setDP2)}
         {tickElement(3, dp3, setDP3)}
         {tickElement(4, dp4, setDP4)}
-        {tickElement(5, dp5, setDP5)}
       </>
     );
   };
@@ -219,7 +216,7 @@ export default function AddDataProduct() {
     const getIcon = () => <AddIcon />;
 
     const enabled = () => {
-      const dp = dp1 || dp2 || dp3 || dp4 || dp5;
+      const dp = dp1 || dp2 || dp3 || dp4;
       return dp && pixelSizeValue > 0 && imageSizeValue > 0;
     };
 
@@ -232,7 +229,7 @@ export default function AddDataProduct() {
           0
         );
       }
-      const observatoryDataProduct = [dp1, dp2, dp3, dp3, dp4];
+      const observatoryDataProduct = [dp1, dp2, dp3, dp4];
       const newDataProduct: DataProductSDP = {
         id: highestId + 1,
         dataProductsSDPId: `SDP-${highestId + 1}`,
@@ -332,6 +329,11 @@ export default function AddDataProduct() {
         </Grid>
         <Grid item xs={3} ml={5}>
           <HelpPanel />
+          <Alert
+            color={AlertColorTypes.Info}
+            text="The associated input options of these observatory data products are under development and subject to change."
+            testId="developmentPanelId"
+          />
         </Grid>
       </Grid>
       {pageFooter()}